Updated inputs for the app
Updated specific timeframe inputs for the first tab, year+quarter+month. Updated input for the supplier in all the tabs, instead of the name we will input the ID in a textbox and not in a drop-down list. For the supplier ID just for example purposes we can write: All, Asus or Acer; in this way it will work the same as the last time with the drop-down list.