Changed inputs of all tabs
Main: Changed label for first tab based on the changed objective of the query.
Repeated Purchases: Yearly timeframe option deleted.
Supplier-Customer Relationship: Start date and end date for the second tab deleted, changed it to select time frame, year, quarter and month.
Time Series: The timeframe will be focused on if it is daily or monthly, being able to specify year, quarter and month for the first one, and specifying the year for the second one.